requirements.txt
# Discord API
discord.py>=2.3.0
discord-py-slash-command>=4.2.1

# HTTP and API
aiohttp>=3.8.4
requests>=2.28.2

# Data processing
numpy>=1.24.3
pandas>=2.0.0

# Visualization
matplotlib>=3.7.1

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.5.0